Split history and location hooks in useRouter
historyとlocationを別のuseEffectに変更
`history.listen` はマウント時 or history 変更時のみ実行してほしいため

ルーティング設定を別ファイルに切り出し
